Split MIR JSON tests
